// src/views/pages/authentication/register-old/types.ts
/**
 **********************************************************************
 * This source code is shared for the sole purpose of demo. Do not copy!
 **********************************************************************
 **/

export type RegisterFormData = {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  submit: string | null;
  referral: string;
};